Create setting_up.py
Copy and paste this line of code in line 3:

print('Hi')

And then press the "Run" button and wait 1-2 seconds.

This should appear in the Terminal window if you did it correctly:

Hi

You are now ready for the journey ahead. (ﾉ ◕ ヮ ◕)ﾉ*:･ﾟ ✧

Press the "Complete" button and then "Next" to continue.

Happy coding!